implement new design with Snap.svg lib
All content in Snap.svg is SVG based. We can finally include GIFs and
animating content works, too. The idea is to use ES6 syntax in the
frontend for a better structure of the code. Each panel / widget is
represented by a separate class.
An enum can then be used to change positions of
panels (LEFT, RIGHT, TOP, BOTTOM).